Override getClaims() method of BasicScopeHandler
In order for the .well-known response to return the set of claims actually
defined in the config file instead of the default set of claims as defined and
returned by BasicScopeHandler.getClaims(), we need to override the latter. For
that we also register the DSOA2ServiceEnvironment instance in the instance of
the DSDynamicScopeHandler, similarly to what is done in the BasicScopeHandler.