Improve transliteration to PrintableString
Move from internal Java Normalizer to ICU-project Transliterator for
transliterating UTF-8 text to ascii. That resolves many of the UTF-8 names
ending up being only XXXX.
Also do punycode encoding in the URLDomainNameFilter, which is the only place
where we need to handle UTF-8 domains. All other organisation attributes are
either schacHomeOrganisation which must be already ascii, or URNs or
organisationDisplayName and those can be again transliterated with the
Transliterator.
Add test for the URLDomainNameFilter.