Fix javamail issue
javax.mail is needed for registration page, but also gets pulled in via
oa4mp-server-admin-oauth2. We change as follows:
- At runtime we only need oa4mp-server-admin-oauth2 for the cli, but at compile
  time we need it for both, hence add as 'scope provided' for the webapp and
  scope compile for the cli.
- Also add manually the mysql-connector-java which would otherwise be pulled in.
Having it runtime in the webapp somehow doesn't work, we still get a
java.lang.NoClassDefFoundError: javax/mail/Authenticator
error, and adding it to /usr/share/java/tomcat it seems both are used and we get
a class cast error. It probably is a clash with another instance already in
tomcat from the javamail rpm which cannot be removed being a depedency of
tomcat.