Fix typo(s)
Fix typo in schacHomeOrganization.
At the same time, we also update the other attributes to more likely candidates.
Note that the latter two will not be used in a setup with a WAYF.
In any case, these defaults will normally not be used since the cfg.xml will
provide the actual config.
This fixes https://github.com/rcauth-eu/aarc-delegation-server/issues/1