Add the sequence_nr to the transactions table
Also add the sequence_nr to the transactions. This makes it easier to match
transactions with trace records.
We also rename everywhere in the code the misleading tracerecord into cnhash
where it concerns the transactions. Unfortunately we cannot do the same with the
table header (see DSOA2TransactionKeys) or we would break backwards
compatibility/upgradeability of the transactions DB.
We use a sequence_nr -1 in case it's not set in the database. Otherwise, it's
unsigned int.