Return JSON error for /getcert endpoint
Since the /getcert endpoint is a backchannel endpoint, intended for services,
it's much better to throw an OA2ATException than a GeneralException to produce a
JSON with an error and error_description. A GeneralException would produce an
HTML page. The clients will parse also the HTML page, but a JSON is certainly
cleaner.